Add readline => false to all swift questions
In specs, switching between using readline and pure ruby forms of input
causes the following IO errors:

    EOFError:
      The input stream is exhausted.

This is because when a readline question is hit, it will pass the IO on
to the c-ext, and it will slurp up all of the remaining input, instead
of only taking a line at a time (which the ruby version does).

There is no functional difference in this change, it just allows our
test suite to set the input all at once, and then have it processed as
needed by the test suite.