Fix input for `say` in specs
There were a few things wrong here:

- `pass` was in the wrong spot
- the `security_protocol` and `api_version` were not question values

I cheated a bit with security_protocol, since it was the default value,
I just used a empty string to allow the newline that would be inputted
as a result to force it to use the default.  This could be changed to be
more explicit and use "3".