Fix prompt validation
Adds a few ": " where needed, and properly reads the full set of prompts
where needed in expect_heard.

This probably changed a bit from what was there previously because of
the change to using `q.readline = false` in a previous commit.  This
might need to be changed or updated if we go a different route.